test/requirements.txt
pyaudio
deepgram-sdk
python-dotenv
